[css-flex] Port some mozilla tests to check-layout
These files, from
external/wpt/css/vendor-imports/mozilla/mozilla-central-reftests/flexbox/flexbox-intrinsic-ratio-00*,
have ~20 test cases each. Chrome fails some of them but not others. With
check-layout we can use -expected.txt files to prevent regressions for
the cases that we pass.  With the current ref tests, we have to disable
an entire file if we fail it even one case in it.

No expectations were changed in the porting. You can see exactly which
lines changed at [1], because it doesn't rename the files.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/2387420

Change-Id: I6fa59df56599c15cccddba4c77afcad6668b9f3e