[scroll-animations] Support element-based offsets in @scroll-timeline
This CL implements support for element-based offsets in the CSS API.
In order to make parsing as painless as possible, the code that
consumes a selector(<id-selector>) has been moved from
at_rule_descriptor_parser to a separate function in css_parsing_utils
(ConsumeSelectorFunction). This makes it possible to reuse that code
for parsing element-based offsets.

Note that the drive-by removal of the ConsumeWhitespace() call at the
beginning of ConsumeScrollOffset is just to align with the convention
that whitespace should be consumed _before_ a Consume* function is
called.

Bug: 1102788, 1074052
Change-Id: I8caa3caa4fe1efe05b4d451c989e10bb9f67267c